+++ client/src/components/NavBar.jsx
@@ -273,7 +273,7 @@ function NavBar() {
               </Link>
             )}
 
-            {typeUser === "users" && (
+            {/* {typeUser === "users" && ( */}
               <Link href={linkToCart} className="group flex items-center justify-center h-full px-5 lg:bg-transparent bg-white hover:bg-white lg:text-white">
                 <span className="material-symbols-rounded group-hover:text-gray-900 text-lg">
                   shopping_cart
@@ -281,7 +281,7 @@ function NavBar() {
                 <span className="ml-2 text-sm font-medium group-hover:text-gray-800">{totalProducts()}</span>
                 <span className="sr-only">items in cart, view bag</span>
               </Link>
-            )}
+            {/* )} */}
           </div>
         </div>
       )}
