+++ client/src/pages/guest/Products.jsx
@@ -5,7 +5,7 @@ import Card from '@/components/card';
 import Filters from '../../components/Filters';
 
 const Products = () => {
-  const id = useSelector(state=>state.userId);
+  const id = useSelector(state=>state.userId)?useSelector(state=>state.userId):"ac5b18b6-6383-4a9f-8e4c-65ad3c93b81a"
   const dispatch = useDispatch();
 
   useEffect(() => {
